Generate cached script plugin loaders on disk
And remove the need for ad-hoc classloaders, use ClassLoaderScopes
instead.

Due to the fact that ClassLoaderScopes need to be locked before loading
classes from them, it is not possible anymore to reuse already loaded
script plugin classes across Projects in a given Build.

This won't be an issue going forward as we move script plugins classes
from the `buildSrc` scope to each `Project` scope in subsequent stories.